requirements.txt
openpyxl>=3.1.2
msoffcrypto-tool>=5.3.1